rewrote httpd package with gorilla/mux
Reason behind this choice is that though httprouter by Julien Schmidt
provides better performance results gorilla's mux allows us to decouple
our code much better. To use httprouter with a well designed API we need
to write an own router type. This gets difficult when testing as we want
to test the handlers with the built-in httptest.