reverted app structure
The app structure which suggests library based development is not
correct in this situation. The packages will never be used outside
this project as well as there will never be any cmd derivates.

Because of this I reverted the app structure back as it was before
this gives us short import paths and feels familiar.