rewrote store package and httpd to store bindings
In store we now use a global mongo session instead of a connection
stored in a struct. Reason for this is that we will never likely
connect to different mongo databases.

Beside of that we now have only one store struct which acts as
mediator between the database backends and the models.

Furthermore we moved the raw models to an own package.